+++ src/uk/co/tolcroft/finance/sheets/SheetAccountType.java
@@ -111,7 +111,7 @@ public class SheetAccountType {
 		Cell    			myTop;
 		Cell    			myBottom;
 		Cell    			myCell;
-		long    			myID;
+		int	    			myID;
 		int     			myCol;
 		int     			myTotal;
 		int					mySteps;
@@ -152,7 +152,7 @@ public class SheetAccountType {
 					 i++) {
 					/* Access the cells by reference */
 					myCell = mySheet.getCell(myCol, i);
-					myID   = Long.parseLong(myCell.getContents());
+					myID   = Integer.parseInt(myCell.getContents());
 					myCell = mySheet.getCell(myCol+1, i);
 				
 					/* Add the value into the finance tables */
@@ -228,7 +228,7 @@ public class SheetAccountType {
 			while ((myCurr  = myIterator.next()) != null) {
 				/* Create the Identifier and Name cells */
 				myCell = new jxl.write.Label(0, myRow, 
-											 Long.toString(myCurr.getId()));
+											 Integer.toString(myCurr.getId()));
 				mySheet.addCell(myCell);
 				
 				/* Add the name to the list */
