+++ src/uk/co/tolcroft/finance/sheets/SheetRate.java
@@ -143,8 +143,8 @@ public class SheetRate {
 		String    		myRate;
 		String    		myBonus;
 		java.util.Date	myEndDate;
-		long      		myID;
-		long      		myAccountID;
+		int      		myID;
+		int      		myAccountID;
 		DateCell  		myDateCell;
 		Cell      		myCell;
 		int       		myTotal;
@@ -187,9 +187,9 @@ public class SheetRate {
 				
 					/* Access id and account id */
 					myCell    	= mySheet.getCell(myCol, i);
-					myID      	= Long.parseLong(myCell.getContents());
+					myID      	= Integer.parseInt(myCell.getContents());
 					myCell    	= mySheet.getCell(myCol+1, i);
-					myAccountID	= Long.parseLong(myCell.getContents());
+					myAccountID	= Integer.parseInt(myCell.getContents());
 				
 					/* Handle Rate */
 					myCell     	= mySheet.getCell(myCol+2, i);
@@ -291,10 +291,10 @@ public class SheetRate {
 			while ((myCurr  = myIterator.next()) != null) {
 				/* Create the Identifier cell */
 				myCell = new jxl.write.Label(0, myRow, 
-											 Long.toString(myCurr.getId()));
+											 Integer.toString(myCurr.getId()));
 				mySheet.addCell(myCell);
 				myCell = new jxl.write.Label(1, myRow,
-											 Long.toString(myCurr.getAccount().getId()));
+											 Integer.toString(myCurr.getAccount().getId()));
 				mySheet.addCell(myCell);
 			
 				/* Create the Rate cells */
